src/app/extensions/cam-quotes/pages/camfil-account-quote-detail-page/components/add-product-dialog/add-product-dialog.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
import { MockComponent } from 'ng-mocks';

import { CoreStoreModule } from 'ish-core/store/core/core-store.module';
import { CamfilCounterComponent } from 'ish-shared/forms/components/camfil-counter/camfil-counter.component';

import { CamQuotesStoreModule } from '../../../../store/cam-quotes-store.module';

import { AddProductDialogComponent } from './add-product-dialog.component';

describe('Add Product Dialog Component', () => {
  let component: AddProductDialogComponent;
  let fixture: ComponentFixture<AddProductDialogComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [AddProductDialogComponent, MockComponent(CamfilCounterComponent)],
      imports: [CamQuotesStoreModule.forTesting(), CoreStoreModule.forTesting()],
      providers: [
        { provide: MAT_DIALOG_DATA, useValue: [] },
        { provide: MatDialogRef, useValue: {} },
      ],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(AddProductDialogComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
